requirements.txt
slack_bolt
slack_sdk
schedule
mysql-connector-python
python-dotenv
